+++ app/src/main/java/com/artemObrazumov/drinkin/cart/presentation/navigation/Routes.kt
@@ -4,9 +4,3 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data object Cart
-
-@Serializable
-data object NewOrder
-
-@Serializable
-data object OrderSuccess
