+++ app/src/main/java/com/artemObrazumov/drinkin/order/data/OrderMockDataSource.kt
@@ -1,15 +1,13 @@
-package com.artemObrazumov.drinkin.cart.data
+package com.artemObrazumov.drinkin.order.data
 
-import com.artemObrazumov.drinkin.cart.domain.data_source.OrderDataSource
-import com.artemObrazumov.drinkin.cart.domain.models.Order
-import com.artemObrazumov.drinkin.cart.domain.models.toProductInOrder
+import com.artemObrazumov.drinkin.order.domain.data_source.OrderDataSource
+import com.artemObrazumov.drinkin.order.domain.models.Order
+import com.artemObrazumov.drinkin.order.domain.models.toProductInOrder
 import com.artemObrazumov.drinkin.cart.domain.usecase.GetProductsInCartFlowUseCase
 import com.artemObrazumov.drinkin.core.domain.util.Error
 import com.artemObrazumov.drinkin.core.domain.util.NetworkError
 import com.artemObrazumov.drinkin.core.domain.util.Result
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.last
-import kotlinx.coroutines.flow.single
 
 class OrderMockDataSource(
     private val getProductsInCartFlowUseCase: GetProductsInCartFlowUseCase
