+++ app/src/main/java/com/artemObrazumov/drinkin/order/domain/usecase/GetOrderUseCase.kt
@@ -1,7 +1,7 @@
-package com.artemObrazumov.drinkin.cart.domain.usecase
+package com.artemObrazumov.drinkin.order.domain.usecase
 
-import com.artemObrazumov.drinkin.cart.domain.data_source.OrderDataSource
-import com.artemObrazumov.drinkin.cart.domain.models.Order
+import com.artemObrazumov.drinkin.order.domain.data_source.OrderDataSource
+import com.artemObrazumov.drinkin.order.domain.models.Order
 import com.artemObrazumov.drinkin.core.domain.util.Error
 import com.artemObrazumov.drinkin.core.domain.util.Result
