+++ app/src/main/java/com/artemObrazumov/drinkin/order/presentation/navigation/Routes.kt
@@ -1,0 +1,9 @@
+package com.artemObrazumov.drinkin.order.presentation.navigation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object NewOrder
+
+@Serializable
+data object OrderSuccess
